Cleanups

+++ source/processor.hpp
@@ -7,5 +7,5 @@ struct Processor {
     virtual void key_on(uint8_t key, uint8_t velocity) {}
     virtual void key_off(uint8_t key) {}
 
-    int32_t ui_panel_index = -1;
+    size_t ui_panel_index = -1;
 };
